requirements.txt
Brotli==1.0.9
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ConfigArgParse==0.15.1
Flask==1.1.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==1.5a2
geventhttpclient==1.5.3
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
locust==2.8.0
locustio==0.13.4
MarkupSafe==1.1.1
msgpack==1.0.3
msgpack-python==0.5.6
odoo-client-lib==1.2.2
OdooLocust==1.2.0
openerp-client-lib==1.1.2
psutil==5.9.0
python-dotenv==0.10.3
pyzmq==18.1.1
requests==2.22.0
roundrobin==0.0.2
six==1.13.0
typing-extensions==4.2.0
urllib3==1.25.7
Werkzeug==0.16.0
